feat: editor store text block implementation cleanup
