+++ apps/web/core/state/editor/data-block.ts
@@ -1,0 +1,9 @@
+// Set a source by id as the source of the data block
+export function setSource() {
+  return;
+}
+
+// Collection -> Space || Space -> Collection
+export function changeSourceType() {
+  return;
+}
