+++ apps/web/core/state/editor/image-block.ts
@@ -1,0 +1,3 @@
+export function setImage() {
+  return;
+}
