+++ apps/web/partials/editor/editor.tsx
@@ -1,13 +1,5 @@
 'use client';
 
-import BulletList from '@tiptap/extension-bullet-list';
-import Document from '@tiptap/extension-document';
-import Gapcursor from '@tiptap/extension-gapcursor';
-import HardBreak from '@tiptap/extension-hard-break';
-import Image from '@tiptap/extension-image';
-import ListItem from '@tiptap/extension-list-item';
-import Placeholder from '@tiptap/extension-placeholder';
-import Text from '@tiptap/extension-text';
 import { EditorContent, Editor as TiptapEditor, useEditor } from '@tiptap/react';
 import cx from 'classnames';
 import { LayoutGroup } from 'framer-motion';
@@ -16,18 +8,14 @@ import * as React from 'react';
 
 import { useUserIsEditing } from '~/core/hooks/use-user-is-editing';
 import { useEditorStore } from '~/core/state/editor/editor-store';
+import { removeIdAttributes } from '~/core/state/editor/utils';
 
 import { Spacer } from '~/design-system/spacer';
 
 import { NoContent } from '../space-tabs/no-content';
-import { ConfiguredCommandExtension } from './command-extension';
-import { removeIdAttributes } from './editor-utils';
-import { HeadingNode } from './heading-node';
+import { tiptapExtensions } from './extensions';
 import { createIdExtension } from './id-extension';
-import { ParagraphNode } from './paragraph-node';
 import { ServerContent } from './server-content';
-import { TableNode } from './table-node';
-import { TrailingNode } from './trailing-node';
 
 interface Props {
   spaceId: string;
@@ -35,55 +23,6 @@ interface Props {
   shouldHandleOwnSpacing?: boolean;
   spacePage?: boolean;
 }
-
-export const tiptapExtensions = [
-  Document,
-  Text,
-  // StarterKit.configure({
-  //   // We're probably only using the Document and Text from the starterkit. Might
-  //   // save us bytes to use it directly instead of through the kit.
-  //   paragraph: false,
-  //   heading: false,
-  //   code: false,
-  //   hardBreak: false,
-  //   gapcursor: false,
-  //   bulletList: false,
-  //   listItem: false,
-  // }),
-  ParagraphNode,
-  HeadingNode,
-  ConfiguredCommandExtension,
-  HardBreak.extend({
-    addKeyboardShortcuts() {
-      // Make hard breaks behave like normal paragraphs
-      const handleEnter = () =>
-        this.editor.commands.first(({ commands }) => [
-          () => commands.newlineInCode(),
-          () => commands.createParagraphNear(),
-          () => commands.liftEmptyBlock(),
-          () => commands.splitBlock(),
-        ]);
-
-      return {
-        Enter: handleEnter,
-        'Mod-Enter': handleEnter,
-        'Shift-Enter': handleEnter,
-      };
-    },
-  }),
-  Gapcursor,
-  TrailingNode,
-  BulletList,
-  ListItem,
-  TableNode,
-  Image,
-  Placeholder.configure({
-    placeholder: ({ node }) => {
-      const isHeading = node.type.name === 'heading';
-      return isHeading ? 'Heading...' : '/ to select content block or write some content...';
-    },
-  }),
-];
 
 export const Editor = React.memo(function Editor({
   shouldHandleOwnSpacing,
@@ -93,7 +32,6 @@ export const Editor = React.memo(function Editor({
 }: Props) {
   const { upsertEditorState, editorJson, blockIds } = useEditorStore();
   const editable = useUserIsEditing(spaceId);
-  const [hasUpdatedEditorJson, setHasUpdatedEditorJson] = React.useState(false);
 
   const extensions = React.useMemo(() => [...tiptapExtensions, createIdExtension(spaceId)], [spaceId]);
 
@@ -111,7 +49,6 @@ export const Editor = React.memo(function Editor({
       // Responsible for converting all editor blocks to triples
       // Fires after the IdExtension's onBlur event which sets the "id" attribute on all nodes
       upsertEditorState(params.editor.getJSON());
-      setHasUpdatedEditorJson(true);
     },
     [upsertEditorState]
   );
